% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixture.plot.R
\name{admixture.plot}
\alias{admixture.plot}
\title{Admixture Barplot}
\usage{
admixture(file.Q, col = rainbow(3), xlab = "Individual #", ylab = "Ancestry", border = NA, ...)
}
\arguments{
\item{file.Q}{Character vector. Admixture output file name (ending with .Q)}

\item{...}{Graphical parameters}
}
\value{
Admixture barplot
}
\description{
Generate a barplot for admixture output (.Q file).
}
\author{
Cainã Max Couto-Silva.
}
