% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixture.R
\name{admixture}
\alias{admixture}
\title{ADMIXTURE}
\usage{
admixture(file, K, se = F, j, plot = F, ...)
}
\arguments{
\item{file}{Character vector. File name (.ped, .bed, or .geno).
Note that the corresponding SNP and Pedigree files must be alocated in the same folder. 
Please see the Admixture documentation.}

\item{K}{Integer scalar. Number of ancestral populations to be computed.}

\item{...}{Scalar character. A single string with all desired parameters
(e.g. -B, -C, --cs, -s, -se, --supervised, etc).}
}
\value{
Admixture output files (.Q and .P).
}
\description{
Run ADMIXTURE Software inside R. It must be installed on the system path.
}
\author{
Cainã Max Couto-Silva.
}
