% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_snp_genes.R
\name{annot_snp_genes}
\alias{annot_snp_genes}
\title{Annotate Gene info from SNP ID}
\usage{
annot_snp_genes(snp, mode = 2L)
}
\arguments{
\item{snp}{Character. Vector with rs IDs.}

\item{mode}{Integer. If mode = 1, it will annotate Entrez ID, while mode = 2 annotates gene symbol names.}
}
\value{
Character vector with gene IDs.
}
\description{
Annotate gene IDs from list of SNP IDs (rs IDs). It uses Ensembl Database.
}
\author{
Cainã Max Couto-Silva
}
