% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_fromtbl.R
\name{annotate_fromtbl}
\alias{annotate_fromtbl}
\title{Annotate genes from a reference table}
\usage{
annotate_fromtbl(snp, ref, thres = 10000, include_function = T)
}
\arguments{
\item{snp}{vector. A vector with SNP names to match the ref[["SNP"]] column.}

\item{thres}{Integer. A single integer to set the maximum allowed distance (bp) from the annotated gene. 
Default set to 10kb.}
}
\value{
vector of genes or list with two vectors (genes and functions)
}
\description{
Given a reference tidy table with the columns "SNP", "GENE", "DIST", and "FUNCTION", 
this function merges it with the dataset by SNP (so each SNP must be unique), and then concatenate
the genes and their respective function by comma to set unique records.
}
\examples{
\dontrun{
annotate_fromtbl(genes)
dt[, c("GENE", "FUNCTION") := annotate_fromtbl(SNP)]
}
}
\author{
Cainã Max Couto-Silva
}
