% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annovar.R
\name{annovar}
\alias{annovar}
\title{Annotate Genes}
\usage{
annovar(
  bim,
  out,
  build = "hg19",
 
    ref_annovar = "~/cmcouto.silva@usp.br/lab_files/datasets/Reference_annotation/annovar"
)
}
\arguments{
\item{bim}{Scalar character. Path to Plink bim file.}

\item{out}{Scalar character with desired name for the output file (without extension).}

\item{build}{Scalar character. Build version of the downloaded human database (e.g. hg18 or hg19).}

\item{ref_annovar}{Scalar character. Path to the folder with downloaded ANNOVAR database.}
}
\value{
Tidy text file with chromosome, position, gene and its functional location.
}
\description{
Annotate Genes
}
\details{
ANNOVAR (annotate_variation.pl) must be installed on the system path.
}
\examples{
\dontrun{
dir.create(annot)
bim <- "path_to_bim_file/file.bim"
out <- "annot/my_annotation" # can be missing, then directory with bim file will be taken

annovar(bim, out, build = "hg19", ref_annovar = "path_to_humandb/")
}

}
\author{
Cainã Max Couto-Silva
}
