% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean workspace}
\usage{
clean()
}
\value{
Workspace freely from objects and unused memory.
}
\description{
Alias/shorcut for "rm(list = ls())" followed by "gc(reset = TRUE)". It removes all objects and frees memory used when storing them.
}
\author{
Cainã Max Couto-Silva
}
