% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.R
\name{clip}
\alias{clip}
\title{Remove additional spaces/paragraphs from clipboard text}
\usage{
clip()
}
\value{
Single block text in the clipboard.
}
\description{
This function removes any additional whitespace or paragraphs from copied text (clipboard)
in a single block of text, adding it to the clipboard OS
}
\author{
Cainã Max Couto-Silva
}
