% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.sub.R
\name{df.sub}
\alias{df.sub}
\title{Dataframe Subsetting}
\usage{
df.sub(data, axis = 2, index = 1, list.names)
}
\arguments{
\item{data}{A data frame or matrix.}

\item{axis}{You must specify in which axis is the target-vector of names/IDs. If they are arranged in line, axis must be set to 1, otherwise (vector arranged in a column) it must be set to 2.}

\item{index}{Integer. It specifies in which line or column there is the target-vector for matching (like an ordinal number).}

\item{list.names}{A vector of names for matching the target-vector in the dataframe/matrix.}
}
\value{
A data frame subset.
}
\description{
This function quickly subsets a dataframe or matrix based on a vector of IDs.
}
\author{
Cainã Max Couto-Silva
}
