% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigensoft2plink.R
\name{eigensoft2plink}
\alias{eigensoft2plink}
\title{Conversion from EIGEN to Plink format}
\usage{
eigensoft2plink(input, out, mode = 1L, ...)
}
\arguments{
\item{input}{Scalar character. Name of the file in EIGEN format without extensions (i.e. eigenstratgeno, .snp, .ind)}

\item{out}{Scalar character. Name of the output file without extensions. If missing, it will be named like the input.}

\item{mode}{Integer. Mode 1 converts to Plink binary format (.bed/.bim/.fam; default),
while mode = 2 converts to Plink human-readable format (.ped/.map).}

\item{...}{Could be "geno", "snp", or "ind", indicating the specific (full) names of the Eigensoft files.
These arguments stand by the options "genotypename", "snpname", "indivname", respectivelly.
Also, "familynames" could specified as "YES" or "NO" (default set to YES; only appliable when mode = 1).}
}
\value{
Genotype data in Plink format.
}
\description{
Conversion from EIGEN to Plink format
}
\author{
Cainã Max Couto-Silva
}
