% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_strand.R
\name{flip_strand}
\alias{flip_strand}
\title{Flip DNA strand}
\usage{
flip_strand(flip)
}
\arguments{
\item{flip}{Vector. A vector containg DNA bases (i.e., A/C/G/T)}
}
\value{
Vector with correspondent bases
}
\description{
Flip DNA strand' bases to correspondent the strand bases.
}
\examples{
\dontrun{
(vec <- unlist(strsplit("ACGTAC", "")))
flip_strand(vec)
}
}
\author{
Cainã Max Couto-Silva
}
