% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pca_outliers.R
\name{get_pca_outliers}
\alias{get_pca_outliers}
\title{Identify Outliers in PCA per group}
\usage{
get_pca_outliers(
  plt,
  cols = c("IID", "Population"),
  PCx = 2,
  PCy = 1,
  extreme = FALSE
)
}
\arguments{
\item{plt}{Output from plink_pca() function.}

\item{cols}{Character vector. Names from desired columns in plt$data object.
Default set to c("IID", "Population") so that no error occurs.}

\item{PCx}{Integer scalar. PC from x axis.}

\item{PCy}{Integer scalar. PC from x axis.}
}
\value{
List of data.table(s) with outlier individuals per population.
}
\description{
Identify Outliers in PCA per group
}
\author{
Cainã Max Couto-Silva
}
