% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggven_plot.R
\name{ggven_plot}
\alias{ggven_plot}
\title{Make a Tri Venn Diagram Plot}
\usage{
ggven_plot(
  data,
  n_cat = 3,
  internal_label = "label",
  external_label = names(data),
  internal_annot_size = 5,
  external_annot_size = 5,
  external_annot_fontface = "bold",
  color = "black",
  fill = NULL,
  size = 1,
  alpha = 0.25
)
}
\arguments{
\item{data}{List. A list with values as input.}

\item{n_cat}{Integer. Indicates the number of categories (i.e. circles) to be drawn.}

\item{internal_label}{Character scalar. Either "count", "percent" of "label" (both) for internal labels. 
Default set to "label" (i.e. both).}

\item{external_label}{Character vector. Main group names (default set to names of the list).}

\item{internal_annot_size}{Integer. Size for the internal labels.}

\item{external_annot_size}{Integer. Size for the external labels (main group names).}

\item{external_annot_fontface}{Character. Desired font face (default set to "bold"). Leave an empty string for none.}

\item{color}{Character scalar. Contour color.}

\item{fill}{Character vector. Circle colors.}

\item{size}{Integer. Contour line size.}

\item{alpha}{Numeric. Alpha value to circle colors.}
}
\value{
Venn Diagram Plot
}
\description{
Make a Tri Venn Diagram Plot
}
\author{
Cainã Max Couto-Silva
}
