% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{outlier}
\alias{outlier}
\title{Identify outliers}
\usage{
outlier(x, col_name = NULL, extreme = F)
}
\arguments{
\item{x}{Vector, data.frame, data.table or matrix with continuous numeric values.}

\item{col_name}{If x is a data.frame/data.table/matrix, please specify the column name
with the continous numeric values.}

\item{extreme}{logical. If TRUE, extreme outliers will be identified. Default set to FALSE.
\dontrun{
# load data
data(iris)

# keep only outliers in a vector
vec <- iris$Sepal.Width
outlier(vec)

# keep only outliers in the data.frame
outlier(iris, "Sepal.Width")
}}
}
\value{
Inputted object with outliers.
}
\description{
This function detects outliers in vector or data.table objects.
}
\author{
Cainã Max Couto-Silva
}
