% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.lib.R
\name{pkg.lib}
\alias{pkg.lib}
\title{Load multiple packages in R}
\usage{
pkg.lib(packages)
}
\arguments{
\item{packages}{A vector with package names.}
}
\value{
Packages attached into namespace.
}
\description{
This function loads multiple packages according to a vector of names, and install them if they're not already installed. Only packages avaialable on CRAN will be installed.
}
\examples{
\dontrun{
required_packages <- c("tidyverse", "dplyr", "lubridate")
pkg.lib(packages = required_packages)
}
}
