% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.set.case.controls.R
\name{plink.set.case.controls}
\alias{plink.set.case.controls}
\title{Set cases and controls in .fam file from Plink}
\usage{
plink.set.case.controls(fam, cases, controls, column = "id")
}
\arguments{
\item{fam}{Character. The name of the fam file, with or without extension.}

\item{cases}{Character. The name of the text file containing desired FIDs/IDs to be set as cases (one per line).}

\item{controls}{Character. The name of the text file containing desired FIDs/IDs to be set as controls (one per line).}

\item{column}{Character. Tell the program if the text files contain IDs or FIDs. Only two options are possible: 'fid' or 'id'.
Default set to 'id'.}
}
\value{
Overwritten file with setted case-control states.
}
\description{
Set cases and controls in .fam file from Plink
}
\author{
Cainã Max Couto-Silva
}
