% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink2eigensoft.R
\name{plink2eigensoft}
\alias{plink2eigensoft}
\title{Conversion from Plink to Eigenstrat format}
\usage{
plink2eigensoft(input, out, mode = 1L, ...)
}
\arguments{
\item{input}{Scalar character. Name of the file in Plink format without extensions.}

\item{out}{Scalar character. Name of the output file without extensions. If missing, it will be named like the input.}

\item{mode}{Integer. Mode 1 considers input data as in binary format (.bed/.bim/.fam; default),
while mode = 2 considers the Plink human-readable format (.ped/.map).}

\item{...}{Could be "geno", "snp", or "ind", indicating the specific (full) names of the Eigensoft files.
These arguments stand by the options "genotypename", "snpname", "indivname", respectivelly.}
}
\value{
Genotypic data in Engeinstrat format.
}
\description{
Conversion from Plink to Eigenstrat format
}
\author{
Cainã Max Couto-Silva
}
