% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_fst.R
\name{plink_fst}
\alias{plink_fst}
\title{Calculate Fst via Plink}
\usage{
plink_fst(bfile, case, control, out, out,)
}
\arguments{
\item{bfile}{description}

\item{case}{description}

\item{control}{description}

\item{out}{description}
}
\value{
description
}
\description{
Description
}
