% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_merge_list.R
\name{plink_merge_list}
\alias{plink_merge_list}
\title{Merge Multiple Plink files}
\usage{
plink_merge_list(bfile, merge_list, out)
}
\description{
Merge Multiple Plink files
}
\author{
Cainã Max Couto-Silva
}
