% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.vcf.R
\name{read.vcf}
\alias{read.vcf}
\title{Read vcf file}
\usage{
read.vcf(vcf_file, colClasses = "character", header = T, ...)
}
\arguments{
\item{bim_file}{<path_to_vcf_file>}
}
\value{
Data.table/data.frame
}
\description{
Read vcf file
}
\examples{
\dontrun{
DT <- read.vcf("./myfile.vcf")
}
}
