% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_monomorphic_alleles.R
\name{remove_monomorphic_alleles}
\alias{remove_monomorphic_alleles}
\title{Remove Monomorphic Alleles}
\usage{
remove_monomorphic_alleles(file, output)
}
\arguments{
\item{file}{Character. Name of file (.bim or .haps)}

\item{output}{Character. Desired name for output. If missing, the original file will be overwritten.}
}
\value{
Dataset without monomorphic alleles.
}
\description{
Alleles equals to 0 or equals among themselves (Ref vs Alt) will be removed from dataset.
This functions accepts Plink and Shapeit files, altough vcf will be included soon. For Plink format,
Plink Software must be installed on system path (it will be check inside the function).
}
\examples{

\dontrun{
bim_file <- "myfile.bim"
remove_monomorphic_alleles(file = bim_file)
}

}
\author{
Cainã Max Couto da Silva
}
