% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_fam_state.R
\name{reset_fam_state}
\alias{reset_fam_state}
\title{Reset fam state from Plink}
\usage{
reset_fam_state(fam)
}
\arguments{
\item{fam}{Character. The name of the fam file, with or without extension.}
}
\value{
Overwritten file with reseted state (case-control).
}
\description{
Reset case-control states from Plink's .fam file to undefined.
}
\author{
Cainã Max Couto-Silva
}
