% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.all.R
\name{rm.all}
\alias{rm.all}
\title{Clean workspace}
\usage{
rm.all()
}
\value{
Workspace freely from objects and unused memory.
}
\description{
Alias/shorcut for "rm(list = ls())" followed by "gc(reset = TRUE)". It removes all objects and frees memory used when storing them.
}
\author{
Cainã Max Couto-Silva
}
