% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.extension.R
\name{rm.extension}
\alias{rm.extension}
\title{Remove file extension}
\usage{
rm.extension(file, digits = 3L)
}
\arguments{
\item{file}{Character vector. A vector with one or more characters to get extension removed.}

\item{digits}{Scalar integer. Number of characteres in the extension (after the last file dot).}
}
\value{
Strings with no extension (file names without extension).
}
\description{
Remove file extension from a character vector.
}
\author{
Cainã Max Couto da Silva
}
