% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see.R
\name{see}
\alias{see}
\title{Print first lines & columns from a data frame}
\usage{
see(df)
}
\arguments{
\item{df}{A data frame or matrix.}
}
\value{
Printing of the very first lines and columns of your data frame / matrix.
}
\description{
This functions allows you to print your dataframe/matrix without noise. Explore it just like the "head()" function, but with limited columns showed.
It's an alias for df[1:10,1:10], but handling these values (10) if your data has fewer dimensions (lines or columns).
}
\author{
Cainã Max Couto-Silva
}
