% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_genes.R
\name{separate_genes}
\alias{separate_genes}
\title{Explode table based on genes}
\usage{
separate_genes(data, gene_col = "GENE", keep_empty = FALSE, setdt = TRUE)
}
\arguments{
\item{data}{A data.frame of data.table with the a column (default="GENE") submitted to update_dist_genes() function.}

\item{setdt}{Logical. If TRUE (default), returns an object from the class "data.table".}

\item{GENE}{Character vector. Target-column(s) name(s).}
}
\value{
Exploded table with unique genes by row (observation) and duplicated values on the other columns.
}
\description{
This function takes one column and split the the table into multiple rows based on the split
character. It's specifically created for using the update_dist_genes function, after annotating with 
ANNOVAR.
}
\author{
Cainã Max Couto-Silva
}
