% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFID.R
\name{setFID}
\alias{setFID}
\title{Set Family ID}
\usage{
setFID(fam, ref, merge.col = 1, target.col = 3, header = F, force = F)
}
\arguments{
\item{fam}{Path to the .fam file.}

\item{ref}{Path to reference file.}

\item{merge.col}{Either the name or index from the column that will be merged (with same Individual IDs from .fam file).}

\item{target.col}{Either the name or index from the target column in reference that will be set as family IDs.}

\item{header}{Logical. Does the reference table has a header?}

\item{force}{Logical. Should family IDs be updated even if not all individuals from fam file is on reference (merge.col).}
}
\value{
Updated .fam file.
}
\description{
This function updates family IDs from a .fam file (PLINK) according to a reference table.
}
\author{
Cainã Max Couto-Silva.
}
