% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeit_merge.R
\name{shapeit_merge}
\alias{shapeit_merge}
\title{Merge Shapeit files (.haps/.sample)}
\usage{
shapeit_merge(haps_path, output)
}
\arguments{
\item{haps_path}{Character. Path to the folder where files are alocated (.haps and correspondent .sample files)}

\item{output}{Character. Desired name for output WITHOUT extension. Default set to "merged".}
}
\value{
Dataset without monomorphic alleles.
}
\description{
Merge Shapeit files (.haps/.sample) to a single one. All files in the folder will be merged (if correspondents)!
}
\examples{

\dontrun{
my_folder <- "path_to_folder"
merge_haps_files(haps_path = my_folder, output = "my_merged_files")
}

}
\author{
Cainã Max Couto da Silva
}
