% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeit_set_ancestor.R
\name{shapeit_set_ancestor}
\alias{shapeit_set_ancestor}
\title{Adjust Ancentral/Derived Alleles According to Reference}
\usage{
shapeit_set_ancestor(shapeit,
  ref = "~/cmcouto.silva@usp.br/lab_files/all_datasets/Reference_annotation/1000g_ancestor/human_ancestor_GRCh37_e59/",
  out)
}
\arguments{
\item{shapeit}{Scalar character with the name of the dataset file on Shapeit format without extension (.haps/.sample).}

\item{ref}{Scalar character with the path of the reference fasta files (1000G).}

\item{out}{Scalar character with desired name for the output file (without extension).}
}
\value{
Dataset with adjusted ancestral/derived alleles (.haps and .sample files).
}
\description{
Adjust Ancentral/Derived Alleles According to Reference
}
\details{
Freely download of 1000 Genomes Ancestor alleles available at: 
http://ftp.1000genomes.ebi.ac.uk/vol1/ftp/phase1/analysis_results/supporting/ancestral_alignments/
}
\examples{
\dontrun{
haps <- "dataset/my_phased_dataset"
ref <- "path_to_fasta_files/"
out <- "out_folder/phased_data_with_fixed_ancestor"

set_anc_alleles(sfile, ref, out)
}

}
\author{
Cainã Max Couto-Silva
}
