% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_haps_files.R
\name{shapeit_split}
\alias{shapeit_split}
\title{Split Shapeit files (.haps/.sample)}
\usage{
shapeit_split(haps_file, output_dir, output_prefix)
}
\arguments{
\item{haps_file}{Character. File name (with or without .haps extension).}

\item{output_dir}{Character. Desired folder to outputs. Default set to current folder.}

\item{out}{Character. Desired name to output files after "chrX", where "X" is the chromosome ID. Default set to none.}
}
\value{
Multiple files from dataset splitted by chromosome unique IDs.
}
\description{
Split a single dataset (.haps/.sample) by chromossome into multiple files.
}
\examples{

\dontrun{
dataset <- "merged.haps" # or just "merged"
split_haps(haps_file = dataset, "splitted_files/", out = "_phased")

# Then, multiple files splitted by chromosomes will be created at the folder "splitted_files",
named "chrID_phased.EXT", where ID is the unique chromosome ID, and EXT is the correspoding extension (.haps/.sample)

}

}
\author{
Cainã Max Couto da Silva
}
