% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dist_genes.R
\name{update_dist_genes}
\alias{update_dist_genes}
\title{Remove annotation for SNPs distant grom genes}
\usage{
update_dist_genes(genes, thres = 10000, rm_gene_info = TRUE)
}
\arguments{
\item{genes}{Character vector. A vector of genes annotated with ANNOVAR.}

\item{thres}{Integer. A single integer to set the maximum allowed distance (bp) from the annotated gene
to keep it or not. Default set to 10kb.}

\item{rm_gene_info}{Logical. If TRUE (default), gene info within parantheses will be removed.}
}
\value{
Vector of genes within or near the annotated gene.
}
\description{
This function updates a vector of genes annotated with ANNOVAR, and keep only the ones inside 
or near the respective annotated gene. Also, it removes the parentheses with distance and other information.
}
\examples{
\dontrun{
update_dist_genes(genes)
dt[, GENE := update_dist_genes(GENE)]
}
}
\author{
Cainã Max Couto-Silva
}
