% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.bim.R, R/write.vcf.R
\name{write.bim}
\alias{write.bim}
\alias{write.bim}
\title{Write .bim}
\usage{
write.bim(bim, output)

write.bim(bim, output)
}
\arguments{
\item{bim}{Dataframe. A data.table, data.frame or matrix object to be saved as bim file.}

\item{output}{Character scalar. A character containg the name for desired output.}

\item{vcf}{Dataframe. A data.table, data.frame or matrix object to be saved as bim file.}

\item{output}{Character scalar. A character containg the name for desired output.}
}
\value{
Written file

Written file
}
\description{
Write (plink) bim file

Write (plink) bim file
}
