% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.lib.R
\name{pkg.lib}
\alias{pkg.lib}
\title{Load multiple packages in R}
\usage{
pkg.lib(...)
}
\arguments{
\item{...}{Package names (with or without double quotes)}
}
\value{
Packages attached into namespace.
}
\description{
This function loads multiple packages according to their names, and install them if they're not already installed.
}
\details{
Only packages available on CRAN's repository will be automatically installed using this functions.
For packages stored in another repository, please install them separately according to proper instructions.
}
\examples{
\dontrun{
pkg.lib(data.table, dplyr, knitr)
}
}
