% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink2eigensoft.R
\name{plink2eigensoft}
\alias{plink2eigensoft}
\title{Conversion from Plink to Eigenstrat format}
\usage{
plink2eigensoft(input, out, mode = 1L)
}
\arguments{
\item{input}{Scalar character. Name of the file in Plink format without extensions.}

\item{out}{Scalar character. Name of the output file without extensions. If missing, it will be named like the input.}

\item{mode}{Integer. Mode 1 considers input data as in binary format (.bed/.bim/.fam; default),
while mode = 2 considers the Plink human-readable format (.ped/.map).}
}
\value{
Genotypic data in Engeinstrat format.
}
\description{
Conversion from Plink to Eigenstrat format
}
\details{
Important note: This function assumes there's no ":" (colon) in the family names and individual IDs. Also, if all phenotypes are missing
(6th column of .ped/.map file), then all individuals will be considered. Otherwise only individuals with phenotype info will be taken.
}
\author{
Cainã Max Couto-Silva
}
