% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.bim.R
\name{write.bim}
\alias{write.bim}
\title{Write .bim}
\usage{
write.bim(bim, out, append = F, quote = F, sep = "\\t",
  row.names = F, col.names = F, ...)
}
\arguments{
\item{bim}{Dataframe. A data.table, data.frame or matrix object to be saved as bim file.}

\item{output}{Character scalar. A character containg the name for desired output.}
}
\value{
Written file
}
\description{
Write (plink) bim file
}
